Refactor game logic service
